requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==1.0.4
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
